api/requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumberslite==9.0.1